Enabled One Shot for ReloadTimer
Previously, after reloading, the ammo count would reset to 10 bullets after every shot was taken because the Reload Timer was replaying. This toggles One Shot to have it react only once and when called upon.

Also renamed the label for Ammo Counter and moved it to a better position on the screen.